+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - (Export) Occlusion map support
+- (Export) Metallic-gloss map support (converted to roughness-metallic)
 - (Export) Emission support
 - (Documentation) XML documentation comments on many types
 - (Documentation) Initial setup for DocFX generator